// src/app/clientes/interfaces/reservaHora.interface.ts
export interface ReservaHora {
    id?:            number;
    nombre?:        string;
    sexo?:          string;
    raza?:          string;
    fechaNacimiento?:string;
    nombreCliente?: string;
    apellidoCliente?:string;
    emailCliente?:  string;
    fechaNacimientoC?:string;
    motivoConsulta?: string;
    veterinario?: string;
    fechaConsulta?: string;
    region?: string;
    comuna?: string;
    sucursal?: string;
}